# Speed up tests