--- a/tests/test_filters_minicircle.py
+++ b/tests/test_filters_minicircle.py
@@ -33,7 +33,7 @@ def test_remove_quadratic_unmasked(minicircle_initial_quadratic_removal: Filters
 def test_get_threshold_otsu(minicircle_threshold_otsu: np.array) -> None:
     """Test calculation of threshold."""
     assert isinstance(minicircle_threshold_otsu.thresholds, dict)
-    assert minicircle_threshold_otsu.thresholds["above"] == pytest.approx(0.40108209618256363)
+    assert minicircle_threshold_otsu.thresholds["above"] == pytest.approx(0.4990958836019106)
 
 
 def test_get_threshold_stddev(minicircle_threshold_stddev: np.array) -> None:
@@ -54,21 +54,21 @@ def test_get_mask(minicircle_mask: Filters) -> None:
     """Test derivation of mask."""
     assert isinstance(minicircle_mask.images["mask"], np.ndarray)
     assert minicircle_mask.images["mask"].shape == (64, 64)
-    assert minicircle_mask.images["mask"].sum() == 616
+    assert minicircle_mask.images["mask"].sum() == 590
 
 
 def test_median_flatten_masked(minicircle_masked_median_flatten: Filters) -> None:
     """Test alignment of rows without mask."""
     assert isinstance(minicircle_masked_median_flatten.images["masked_median_flatten"], np.ndarray)
     assert minicircle_masked_median_flatten.images["masked_median_flatten"].shape == (64, 64)
-    assert minicircle_masked_median_flatten.images["masked_median_flatten"].sum() == pytest.approx(1348.320244358951)
+    assert minicircle_masked_median_flatten.images["masked_median_flatten"].sum() == pytest.approx(1341.2916822670288)
 
 
 def test_remove_x_y_tilt_masked(minicircle_masked_tilt_removal: Filters) -> None:
     """Test removal of tilt with mask."""
     assert isinstance(minicircle_masked_tilt_removal.images["masked_tilt_removal"], np.ndarray)
     assert minicircle_masked_tilt_removal.images["masked_tilt_removal"].shape == (64, 64)
-    assert minicircle_masked_tilt_removal.images["masked_tilt_removal"].sum() == pytest.approx(1339.6888686354937)
+    assert minicircle_masked_tilt_removal.images["masked_tilt_removal"].sum() == pytest.approx(1324.5046829300152)
 
 
 def test_remove_quadratic_masked(minicircle_masked_quadratic_removal: Filters) -> None:
@@ -76,7 +76,7 @@ def test_remove_quadratic_masked(minicircle_masked_quadratic_removal: Filters) -
     assert isinstance(minicircle_masked_quadratic_removal.images["masked_quadratic_removal"], np.ndarray)
     assert minicircle_masked_quadratic_removal.images["masked_quadratic_removal"].shape == (64, 64)
     assert minicircle_masked_quadratic_removal.images["masked_quadratic_removal"].sum() == pytest.approx(
-        1336.8758836167635
+        1324.635491325745
     )
 
 
@@ -84,4 +84,4 @@ def test_gaussian_filter(minicircle_grain_gaussian_filter: Filters) -> None:
     """Test gaussian filter applied to background."""
     assert isinstance(minicircle_grain_gaussian_filter.images["gaussian_filtered"], np.ndarray)
     assert minicircle_grain_gaussian_filter.images["gaussian_filtered"].shape == (64, 64)
-    assert minicircle_grain_gaussian_filter.images["gaussian_filtered"].sum() == pytest.approx(1338.0528938158313)
+    assert minicircle_grain_gaussian_filter.images["gaussian_filtered"].sum() == pytest.approx(1325.8136941472153)