--- a/tests/test_io.py
+++ b/tests/test_io.py
@@ -431,6 +431,8 @@ def test_gwy_read_component(load_scan_dummy: LoadScans) -> None:
         ("load_scan_spm", 1, (1024, 1024), 30695369.188316286, "minicircle", 0.4940029296875),
         ("load_scan_ibw", 1, (512, 512), -218091520.0, "minicircle2", 1.5625),
         ("load_scan_jpk", 1, (256, 256), 286598232.9308627, "file", 1.2770176335964876),
+        ("load_scan_gwy", 1, (512, 512), 33836850.232917726, "file", 0.8468632812499975),
+        ("load_scan_topostats", 1, (1024, 1024), 182067.12616107278, "file", 0.4940029296875),
     ],
 )
 def test_load_scan_get_data(