Closes #508 

This PR is aimed at increasing the speed of tests. It does this by replacing the `(1024, 1024)` `minicircle.spm` testing image, with a slice of this image, namely `minicircle[-296:,-296:]` and compressed down from `(256, 256)` to `(64, 64)`. This image is saved in the ✨`.topostats`✨  file format instead of `.spm`.


I have checked that it flattens appropriately, grains are detected, and traces are produced despite this low resolution. (It also looks really cute).
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/53116647-6324-44ad-9a6e-3990cbaba837)
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/b3ff4408-792c-46ca-a5e8-0e19618df0a2)
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/e1ffb578-6aa6-4f39-bb56-b93f232bb4af)

## Optimisation

Previously, running the automated test suite on GitHub would take up to 46 minutes:
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/7d863ffb-cf5b-434e-a807-6fa7cfc544d7)

With this PR, the tests now take at maximum, 7 minutes:
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/b99a3577-64ce-46a1-a1b3-54a4515d4575)

Which is a 6.5x speed increase (or 85% decrease in time taken).

Previously running locally on my Macbook with one core took 612 seconds:
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/2397f973-0cc4-420a-8f9b-3e87a1dc8562)

Now, running the tests locally on my Macbook with one core, it takes 55 seconds:
![image](https://github.com/AFM-SPM/TopoStats/assets/86117496/d63ad9fb-737c-4885-ad12-b5f0355f99d3)


The bulk of the testing time is taken up by (courtesy of the `--duration=0` `pytest` flag):
```
============= slowest durations ==============
7.71s call     tests/test_processing.py::test_save_format[png]
7.66s call     tests/test_processing.py::test_save_cropped_grains[all-True]
5.74s call     tests/test_processing.py::test_save_format[tif]
4.31s call     tests/test_run_topostats.py::test_run_topostats_process_debug
3.86s call     tests/test_run_topostats.py::test_run_topostats_process_all
1.88s call     tests/test_plottingfuncs.py::test_high_dpi
0.94s call     tests/test_io.py::test_load_scan_gwy
```

If we look at the timings from before
```
============= slowest durations ===============
50.61s call     tests/test_processing.py::test_save_format[png]
49.70s call     tests/test_processing.py::test_save_cropped_grains[all-True]
45.45s call     tests/test_processing.py::test_save_format[tif]
39.00s call     tests/test_run_topostats.py::test_run_topostats_process_all
38.41s call     tests/test_run_topostats.py::test_run_topostats_process_debug
37.48s call     tests/test_processing.py::test_process_scan_both
35.82s call     tests/test_processing.py::test_process_stages[True-True-True-True-Traced grain 21 of 21-Combining ['above'] grain statistics and dnatracing statistics]
34.78s call     tests/test_processing.py::test_process_scan_above
34.26s call     tests/test_processing.py::test_save_cropped_grains[core-False]
33.38s call     tests/test_processing.py::test_process_scan_below
32.57s call     tests/test_processing.py::test_process_stages[True-True-True-False-Processing grain-Calculation of DNA Tracing disabled, returning grainstats data frame.]
32.02s call     tests/test_processing.py::test_process_stages[True-True-False-False-Calculation of grainstats disabled, returning empty dataframe.-24-labelled_image_bboxes]
31.69s call     tests/test_processing.py::test_process_stages[True-False-False-False-Detection of grains disabled, returning empty data frame.-15-gaussian_filtered]
30.87s call     tests/test_processing.py::test_run_filters
6.45s call     tests/test_processing.py::test_process_scan_align_grainstats_dnatracing
3.29s call     tests/test_processing.py::test_process_scan_no_grains
```

The culprits are the same, but we have significantly reduced their length in this PR.

If we want further optimisation, we could consider dropping `test_processing.py::test_save_format` which would shorten the test timing by a further 25% (55s to 40s locally), and I'm not sure that testing the image output file format by running `processing` is entirely necessary, however at this point further optimisation might be silly. Thoughts appreciated.



## A note on test validity.
Given that I have changed the main testing image from the full `minicircle.spm` to a cropped, compressed 64x64 region of `minicircle.spm`, I wanted to make sure that the tests were all legitimate and made sense, as if I blindly forced all the tests to succeed, I might be inadvertently ruining some of the tests.

To prevent this, I plotted every image for every test to make sure that each test does what it should do. An example of where this was necessary is in `tests/test_grains_minicircle.py::test_remove_small_objects` where I found that the new image had not got any small grains to remove, and so the test verified nothing, despite passing. I thus updated the height thresholds and size thresholds to allow the proteins through so that then the test can verify that they are removed.

I believe I have ensured that all the changes to the tests do not prevent them from testing what they were intended to test, but if you are skeptical then please do ask. It's better that I spend time now verifying that tests are still functioning (not passing, but testing what they should) than to find out later down the line that things are broken 👍


## Other notes

- Fixed several type-hints that were inaccurate in `conftest.py`.
- TopoSum, when run after `process_scan`, no longer takes the command-line arguments. This is essential, as if you pass the flag `--file_ext` for `process_scan`, as, say, `.spm`, then TopoSum would also take that as the file extension to save images as, and config validation would fail. So now TopoSum will only take command line arguments when run directly.