@ns-rse 2023-07-28
Closes #508 

(Not sure if that works from comments but lets find out!).